fix VS code behind generation (#855)
* fix redirects
* fix serialization problem with Visual Studio
* fix lineending
