+++ TechTalk.SpecFlow.Generator/Configuration/IGeneratorConfigurationProvider.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using TechTalk.SpecFlow.Configuration;
+using TechTalk.SpecFlow.Generator.Interfaces;
 using TechTalk.SpecFlow.Plugins;
 
 namespace TechTalk.SpecFlow.Generator.Configuration
