+++ TechTalk.SpecFlow.Tools/config.template
@@ -4,19 +4,19 @@
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
         <assemblyIdentity name="TechTalk.SpecFlow.Generator" publicKeyToken="0778194805d6db41" culture="neutral" />
-        <bindingRedirect oldVersion="2.0.0.0-[CURRENT_VERSION].0" newVersion="[CURRENT_VERSION].0" />
+        <bindingRedirect oldVersion="1.9.0.0-[CURRENT_VERSION].0" newVersion="[CURRENT_VERSION].0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="TechTalk.SpecFlow.Parser" publicKeyToken="0778194805d6db41" culture="neutral" />
-        <bindingRedirect oldVersion="2.0.0.0-[CURRENT_VERSION].0" newVersion="[CURRENT_VERSION].0" />
+        <bindingRedirect oldVersion="1.9.0.0-[CURRENT_VERSION].0" newVersion="[CURRENT_VERSION].0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="TechTalk.SpecFlow.Utils" publicKeyToken="0778194805d6db41" culture="neutral" />
-        <bindingRedirect oldVersion="2.0.0.0-[CURRENT_VERSION].0" newVersion="[CURRENT_VERSION].0" />
+        <bindingRedirect oldVersion="1.9.0.0-[CURRENT_VERSION].0" newVersion="[CURRENT_VERSION].0" />
       </dependentAssembly>
       <dependentAssembly>
         <assemblyIdentity name="TechTalk.SpecFlow" publicKeyToken="0778194805d6db41" culture="neutral" />
-        <bindingRedirect oldVersion="2.0.0.0-[CURRENT_VERSION].0" newVersion="[CURRENT_VERSION].0" />
+        <bindingRedirect oldVersion="1.9.0.0-[CURRENT_VERSION].0" newVersion="[CURRENT_VERSION].0" />
       </dependentAssembly>
     </assemblyBinding>
   </runtime>
