+++ TechTalk.SpecFlow/Configuration/ISpecFlowConfigurationHolder.cs
@@ -1,0 +1,9 @@
+﻿namespace TechTalk.SpecFlow.Configuration
+{
+    public interface ISpecFlowConfigurationHolder
+    {
+        ConfigSource ConfigSource { get; }
+        string Content { get; }
+        bool HasConfiguration { get; }
+    }
+}
